+++ delphin/tsdb.py
@@ -14,7 +14,7 @@ from collections import OrderedDict
 from gzip import open as gzopen
 import tempfile
 import shutil
-from datetime import datetime
+from datetime import datetime, date
 import warnings
 
 from delphin.exceptions import PyDelphinException, PyDelphinWarning
@@ -64,7 +64,7 @@ _MONTHS = {
 
 RawValue = Union[str, None]
 RawRecord = Sequence[RawValue]
-Value = Union[str, int, float, datetime, None]
+Value = Union[str, int, float, datetime, date, None]
 Record = Sequence[Value]
 Records = Iterable[Record]
 ColumnMap = Dict[str, Value]  # e.g., a partial Record
@@ -680,10 +680,11 @@ def format(datatype: str,
         else:
             default = str(default)  # ensure it is a string
         raw_value = default
-    elif datatype == ':date' and isinstance(value, datetime):
+    elif datatype == ':date' and isinstance(value, (date, datetime)):
         month = _MONTHS[value.month]
         pattern = f'{value.day!s}-{month}-%Y'
-        if (value.hour, value.minute, value.second) != (0, 0, 0):
+        if (isinstance(value, datetime)
+                and (value.hour, value.minute, value.second) != (0, 0, 0)):
             pattern += ' %H:%M:%S'
         raw_value = value.strftime(pattern)
     else:
@@ -774,7 +775,7 @@ def open(dir: util.PathLike,
 def write(dir: util.PathLike,
           name: str,
           records: Iterable[Record],
-          fields: Fields,
+          fields: Fields = None,
           append: bool = False,
           gzip: bool = False,
           encoding: str = 'utf-8') -> None:
@@ -815,7 +816,8 @@ def write(dir: util.PathLike,
         dir: path to the database directory
         name: name of the relation to write
         records: iterable of records to write
-        fields: iterable of :class:`Field` objects
+        fields: iterable of :class:`Field` objects, optional if *dir*
+            points to an existing test suite directory
         append: if `True`, append to rather than overwrite the file
         gzip: if `True` and the file is not empty, compress the file
             with `gzip`; if `False`, do not compress
@@ -833,6 +835,14 @@ def write(dir: util.PathLike,
 
     if not dir.is_dir():
         raise TSDBError(f'invalid test suite directory: {dir}')
+
+    if fields is None:
+        schema_path = dir / SCHEMA_FILENAME
+        if schema_path.is_file():
+            fields = read_schema(schema_path)[name]
+        else:
+            raise TSDBError(
+                f'cannot determine fields; no schema file at {schema_path}')
 
     tx_path, gz_path, use_gz = _get_paths(dir, name)
     if append and (gzip or use_gz):